Updated BOM and added PDF Version
It was noticed that I had the wrong number of linear rails and ball screw. That has been corrected and some notes on the BOM. It was also noted that I did not include the cost of the ATP-5 plates. I have added in the raw materials for the aluminum for fabricated parts that I was missing. If someone wanted to build the design they would still need to fabricate the plates which is one of the biggest hurdles to being able to make this machine. I know this, but it is limitation to any DIY from scratch machine.

I want to thank people for pointing out these issues and missing information. I want this design to be as comprehensive as possible. Again I don't expect anyone will actually replicate it, or if they do they will customize it to fit their needs. But any information I can add will help others when they take on a similar project themselves.